Make create user form and its respective reducers
The leads folder was renamed to createUser and imports and calls
were fixed to match this change, also every reducer called Leads or
similar was changed to Users to be more specific.

Some minor changes were included in this commit as the navbar name and
the title of the web page.