Show error alerts when registration fails
The provider for the react alerts was added to App.js and it was
added an action to get the error messages that are catched
when the ADD_USER action fails.

The alert message is obtained from the server, we need to confirm
that the error in the form don't alter the database sequences, also
regex validations should be added too.